fix(type): SchemaTooNew version u64 → u32 (match schema_version field type)

+++ src-tauri/src/model.rs
@@ -398,7 +398,7 @@ pub enum ConfigState {
     /// Never constructed by the current loader; kept as the forward
     /// contract for handling future config schema versions.
     #[allow(dead_code)]
-    SchemaTooNew { version: u64 },
+    SchemaTooNew { version: u32 },
 }
 
 impl ConfigState {
